+++ libraries/config/src/main/java/net/ornithemc/osl/config/api/LoadingPhase.java
@@ -1,5 +1,8 @@
 package net.ornithemc.osl.config.api;
 
+/**
+ * Different phases in which a scope will load its configs.
+ */
 public enum LoadingPhase {
 
 	START, READY
