+++ libraries/config/src/main/java/net/ornithemc/osl/config/api/config/option/Option.java
@@ -8,4 +8,8 @@ public interface Option {
 
 	void reset();
 
+	void load();
+
+	void unload();
+
 }
