+++ libraries/config/src/main/java/net/ornithemc/osl/config/impl/ConfigManagerImpl.java
@@ -1,115 +1,62 @@
 package net.ornithemc.osl.config.impl;
 
-import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Collection;
-import java.util.LinkedHashMap;
+import java.util.EnumMap;
 import java.util.Map;
+import java.util.function.Consumer;
 
-import net.ornithemc.osl.config.api.config.Config;
-import net.ornithemc.osl.config.api.serdes.FileSerializerType;
-import net.ornithemc.osl.config.api.serdes.SerializationSettings;
-import net.ornithemc.osl.config.api.serdes.config.ConfigSerializer;
-import net.ornithemc.osl.config.api.serdes.config.ConfigSerializers;
 import net.ornithemc.osl.config.api.ConfigScope;
 import net.ornithemc.osl.config.api.LoadingPhase;
+import net.ornithemc.osl.config.api.config.Config;
 
 public final class ConfigManagerImpl {
 
-	private final ConfigScope scope;
-	private final Path rootDir;
-
-	private final Map<String, Config> configs;
-
-	public ConfigManagerImpl(ConfigScope scope, Path rootDir) {
-		this.scope = scope;
-		this.rootDir = rootDir;
-
-		this.configs = new LinkedHashMap<>();
+	public static void register(Config config) {
+		forInstance(config.getScope(), manager -> manager.register(config));
 	}
 
-	public ConfigScope getScope() {
-		return scope;
+	public static void load(ConfigScope scope, LoadingPhase phase) {
+		forInstance(scope, manager -> manager.load(phase));
 	}
 
-	public Path getConfigDir() {
-		return rootDir.resolve(scope.getPath());
+	public static void unload(ConfigScope scope) {
+		forInstance(scope, manager -> manager.unload());
 	}
 
-	public void register(Config config) {
-		configs.compute(key(config.getNamespace(), config.getName()), (key, value) -> {
-			if (value != null) {
-				if (value == config) {
-					return value; // odd thing to register a config multiple times, but okay
-				} else {
-					throw new IllegalStateException("config " + key + " is not unique");
+	public static void save(ConfigScope scope) {
+		forInstance(scope, manager -> manager.save());
+	}
+
+	private static void forInstance(ConfigScope scope, Consumer<ScopedConfigManager> action) {
+		ScopedConfigManager manager = MANAGERS.get(scope);
+
+		if (manager == null) {
+			InternalConfigEvents.CONFIG_MANAGER_READY.register(readyScope -> {
+				if (readyScope == scope) {
+					forInstance(scope, action);
 				}
-			}
-
-			return config;
-		});
-	}
-
-	private static String key(String namespace, String name) {
-		return namespace == null ? name : (namespace + ":" + name);
-	}
-
-	public Collection<Config> getConfigs() {
-		return configs.values();
-	}
-
-	public void load(LoadingPhase phase) {
-		for (Config config : getConfigs()) {
-			if (config.getLoadingPhase() == phase) {
-				config.load();
-			}
-		}
-	}
-
-	public void unload() {
-		for (Config config : getConfigs()) {
-			config.unload();
-		}
-	}
-
-	public void save() {
-		SerializationSettings settings = new SerializationSettings(); // TODO
-
-		for (Config config : getConfigs()) {
-			FileSerializerType<?> type = config.getType();
-
-			try {
-				save(config, settings, type);
-			} catch (IOException e) {
-				// TODO
-			}
-		}
-	}
-
-	private <M> void save(Config config, SerializationSettings settings, FileSerializerType<M> type) throws IOException {
-		ConfigSerializer<M> serializer = ConfigSerializers.get(type);
-
-		if (serializer == null) {
-			throw new IOException("don't know how to serialize " + config + " to file!");
+			});
 		} else {
-			Path path = getFilePath(config);
-			M medium = type.open(path);
-
-			if (!Files.exists(path)) {
-				Files.createDirectories(path.getParent());
-			}
-
-			serializer.serialize(config, settings, medium);
+			action.accept(manager);
 		}
 	}
 
-	private Path getFilePath(Config config) {
-		Path dir = getConfigDir();
+	private static final Map<ConfigScope, ScopedConfigManager> MANAGERS = new EnumMap<>(ConfigScope.class);
 
-		String namespace = config.getNamespace();
-		String name = config.getSaveName();
+	public static void setUp(ConfigScope scope, Path rootDir) {
+		if (MANAGERS.containsKey(scope)) {
+			throw new IllegalStateException("tried to set up a config manager for the " + scope.name().toLowerCase() + " scope while it was already set up!");
+		}
 
-		return (namespace == null) ? dir.resolve(name) :  dir.resolve(namespace).resolve(name);
+		MANAGERS.put(scope, new ScopedConfigManager(scope, rootDir));
+		InternalConfigEvents.CONFIG_MANAGER_READY.invoker().accept(scope);
+	}
+
+	public static void destroy(ConfigScope scope) {
+		if (!MANAGERS.containsKey(scope)) {
+			throw new IllegalStateException("tried to destroy the config manager for the " + scope.name().toLowerCase() + " scope while it was not set up!");
+		}
+
+		MANAGERS.remove(scope);
 	}
 }
